use mixin class instead
