+++ src/scipp/plotting/figure3d.py
@@ -2,8 +2,9 @@
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 # @author Neil Vaytet
 from .. import config
+from .displayable import Displayable
 from .toolbar import PlotToolbar3d
-from .tools import fig_to_pngbytes, widgets_have_mimebundle
+from .tools import fig_to_pngbytes
 from ..utils import value_to_string
 import numpy as np
 import ipywidgets as ipw
@@ -15,7 +16,7 @@ import pythreejs as p3
 from copy import copy
 
 
-class PlotFigure3d:
+class PlotFigure3d(Displayable):
     """
     Class for 3 dimensional plots.
 
@@ -95,21 +96,6 @@ class PlotFigure3d:
 
     def initialize_toolbar(self, **kwargs):
         self.toolbar.initialize(**kwargs)
-
-    if widgets_have_mimebundle():
-
-        def _repr_mimebundle_(self, include=None, exclude=None):
-            """
-            Mimebundle display representation for jupyter notebooks.
-            """
-            return self._to_widget()._repr_mimebundle_(include=include, exclude=exclude)
-    else:
-
-        def _ipython_display_(self):
-            """
-            IPython display representation for Jupyter notebooks.
-            """
-            return self._to_widget()._ipython_display_()
 
     def _to_widget(self):
         """
