+++ src/scipp/plotting/panel.py
@@ -3,10 +3,10 @@
 # @author Neil Vaytet
 
 import ipywidgets as ipw
-from .tools import widgets_have_mimebundle
+from .displayable import Displayable
 
 
-class PlotPanel:
+class PlotPanel(Displayable):
     """
     Base class for providing additional widgets on top of the base dimension
     sliders and mask display control.
@@ -15,21 +15,6 @@ class PlotPanel:
     def __init__(self):
         self.container = ipw.VBox()
         self.controller = None
-
-    if widgets_have_mimebundle():
-
-        def _repr_mimebundle_(self, include=None, exclude=None):
-            """
-            Mimebundle display representation for jupyter notebooks.
-            """
-            return self._to_widget()._repr_mimebundle_(include=include, exclude=exclude)
-    else:
-
-        def _ipython_display_(self):
-            """
-            IPython display representation for Jupyter notebooks.
-            """
-            return self._to_widget()._ipython_display_()
 
     def _to_widget(self):
         """
